requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-bootstrap4==3.0.1
django-classy-tags==2.0.0
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-stdimage==5.3.0
django-treebeard==4.5.1
easy-thumbnails==2.7.1
gunicorn==20.1.0
html5lib==1.1
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
python-decouple==3.6
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
tzlocal==2.1
Unidecode==1.1.2
webencodings==0.5.1
